# teachify-backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Teachify Backend","description":"AI-powered lecture video generation backend for Teachify.","version":"1.0.0"},"paths":{"/auth/register":{"post":{"tags":["Authentication"],"summary":"Register","description":"Register a new user with unique username.","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPublic"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/refresh":{"post":{"tags":["Authentication"],"summary":"Refresh Token","operationId":"refresh_token_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/logout":{"post":{"tags":["Authentication"],"summary":"Logout","operationId":"logout_auth_logout_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/profile":{"get":{"tags":["Authentication"],"summary":"Read Users Me","operationId":"read_users_me_auth_profile_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPublic"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/v1/content/generate":{"post":{"tags":["Lecture Generation"],"summary":"Generate Lecture","description":"Generate AI-based lecture content, visuals, and avatar video (frontend handles compilation).","operationId":"generate_lecture_v1_content_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerationRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LectureOutput"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/v1/content/generate-from-docs":{"post":{"tags":["Lecture Generation"],"summary":"Generate Lecture From Docs","description":"Generate lecture grounded in uploaded documents (RAG). Frontend compiles final video.","operationId":"generate_lecture_from_docs_v1_content_generate_from_docs_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_generate_lecture_from_docs_v1_content_generate_from_docs_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LectureOutput"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_generate_lecture_from_docs_v1_content_generate_from_docs_post":{"properties":{"prompt":{"type":"string","minLength":10,"title":"Prompt"},"files":{"items":{"type":"string","format":"binary"},"type":"array","title":"Files"}},"type":"object","required":["prompt","files"],"title":"Body_generate_lecture_from_docs_v1_content_generate_from_docs_post"},"Body_login_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_auth_login_post"},"GenerationRequest":{"properties":{"prompt":{"type":"string","minLength":10,"title":"Prompt","description":"User topic or question to generate lecture content."}},"type":"object","required":["prompt"],"title":"GenerationRequest","description":"Input schema for content generation endpoint."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LectureOutput":{"properties":{"topic":{"type":"string","title":"Topic"},"introduction":{"type":"string","title":"Introduction"},"main_body":{"type":"string","title":"Main Body"},"conclusion":{"type":"string","title":"Conclusion"},"visualizations":{"items":{"$ref":"#/components/schemas/VisualizationPrompt"},"type":"array","title":"Visualizations"},"video_path":{"type":"string","title":"Video Path","description":"Path or URL to final compiled video file."}},"type":"object","required":["topic","introduction","main_body","conclusion","visualizations","video_path"],"title":"LectureOutput","description":"Final response after full video generation."},"RefreshIn":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token"}},"type":"object","required":["refresh_token"],"title":"RefreshIn"},"UserCreate":{"properties":{"username":{"type":"string","maxLength":50,"minLength":3,"title":"Username"},"email":{"type":"string","format":"email","title":"Email","description":"User email (must be unique)"},"company":{"type":"string","maxLength":128,"title":"Company","default":""},"password":{"type":"string","minLength":8,"title":"Password","description":"User password (will be hashed before saving)."}},"type":"object","required":["username","email","password"],"title":"UserCreate","description":"Schema for new user registration."},"UserPublic":{"properties":{"username":{"type":"string","title":"Username"},"company":{"type":"string","title":"Company"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["username","company","created_at"],"title":"UserPublic","description":"Public model for user responses (no password exposure)."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VisualizationPrompt":{"properties":{"section":{"type":"string","title":"Section","description":"Lecture section this visual relates to (e.g., 'introduction')."},"prompt":{"type":"string","title":"Prompt","description":"Detailed image generation prompt."},"image_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Path","description":"Path or URL to generated image."},"paragraph_index":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Paragraph Index","description":"Index of the paragraph in that section."},"snippet":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Snippet","description":"Exact text snippet where this image applies."}},"type":"object","required":["section","prompt"],"title":"VisualizationPrompt","description":"Model representing one image prompt for a lecture section."}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/login"}}}}}}